Add files via upload
This is an Implementation of a binary search tree in java. It also has the implementation of a lazy deletion in the binary search tree, where once you delete a node instead of deleting the node is marked deleted. The actual deleting only happens when 30% of all the nodes are marked. This is to compare the running time of this two implementantions of the binary search tree. There is actually also use of Java clock to time the run time of the two algorithms. There is also JUnit tests for the classes created.